tests: refactor solomachine tests to use light client module as entrypoint. (#6038)
* tests: refactor tests to use light client module as entrypoint.

* Rename solomachine clients identifiers.

* ref: clean up previous tests, add two more for LatestHeight, UpdateStatus.

* ref: add verifying funcs.

* rm: old tests.

* nits: test names

* Apply suggestions from code review


* Update modules/light-clients/06-solomachine/solomachine_test.go


* chore: rm one of the consts

* review: address feedback.

---------
