+++ modules/light-clients/06-solomachine/proof.go
@@ -26,7 +26,7 @@ func VerifySignature(pubKey cryptotypes.PubKey, signBytes []byte, sigData signin
 		if err := pubKey.VerifyMultisignature(func(signing.SignMode) ([]byte, error) {
 			return signBytes, nil
 		}, data); err != nil {
-			return err
+			return errorsmod.Wrapf(ErrSignatureVerificationFailed, err.Error())
 		}
 
 	default:
